Recombine ManagedWebSocket.netcoreapp.cs into ManagedWebSocket.cs
